+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog for package simple_serial_port_reader
 ------------------
 * C++11
 * Support start/stop commands
+* Rename param 'debug' to 'verbose'
 * README
 
 0.0.1 (2021-03-03)